refactor power off reboot testcase
